Login and Auth, Flash msg and Delete Auth, Search and Goodbye and my skrinshots

+++ Jobseeker/App/Controllers/ListingController.php
@@ -4,7 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
-
+use Framework\Session;
+use Framework\Authorization;
 
 
 class ListingController
@@ -20,7 +21,7 @@ class ListingController
     public function index()
     {
 
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', ['listings' => $listings]);
     }
@@ -58,7 +59,7 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id']; 
+        $newListingData['user_id'] = Session::get('user')['id'];; // Assuming user_id is 1 for demonstration
         $newListingData = array_map('sanitize', $newListingData);
 
         $requiredFields = ['title', 'description', 'salary', 'email', 'city', 'state'];
@@ -126,6 +127,12 @@ class ListingController
             ErrorController::notFound('Listing not found');
             return;
         }
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+            return redirect('/listings/' . $listing->id);
+        }
+
 
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
@@ -144,11 +151,18 @@ class ListingController
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
-        //check if listing exists
+        //check if listig exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
         }
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+            return redirect('/listings/' . $listing->id);
+        }
+
+        loadView('listings/edit', ['listing' => $listing]);
     }
     /**
      * Update a listing
@@ -172,6 +186,11 @@ class ListingController
             return;
         }
 
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+            return redirect('/listings/' . $listing->id);
+        }
 
         $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'email', 'requirements', 'benefits'];
 
@@ -195,6 +214,7 @@ class ListingController
                 'errors' => $errors
             ]);
         } else {
+            
             //submit to database
             $updateFields = [];
 
@@ -207,6 +227,10 @@ class ListingController
 
             $updatedValues['id'] = $id;
             $this->db->query($updateQuery, $updatedValues);
+
+            Session::setFlashMessage('success_message', 'Listing successfully updated');
+
+            redirect('/listings/' . $id);
         }
     }
     /**
@@ -214,5 +238,24 @@ class ListingController
      *
      * @return void
      */
-   
+    public function search()
+    {
+        $keywords = isset($_GET['keywords']) ? trim($_GET['keywords']) : '';
+        $location = isset($_GET['location']) ? trim($_GET['location']) : '';
+
+        $query = "SELECT * FROM listings WHERE (title LIKE :keywords OR description LIKE :keywords OR tags LIKE :keywords OR company LIKE :keywords) AND (city LIKE :location OR state LIKE :location)";
+
+        $params = [
+            'keywords' => "%{$keywords}%",
+            'location' => "%{$location}%"
+        ];
+
+        $listings = $this->db->query($query, $params)->fetchAll();
+
+        loadView('/listings/index', [
+            'listings' => $listings,
+            'keywords' => $keywords,
+            'location' => $location
+        ]);
+    }
 }
